Kollision abermals verbessert

+++ core/src/de/uniwuerzburg/battletanks/Entity.java
@@ -83,23 +83,23 @@ public class Entity {
         this.speed = speed;
     }
 
-    public int getX() {
-        return (int) position.x;
+    public float getX() {
+        return position.x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         position.x = x;
     }
 
-    public int getY() {
-        return (int) position.y;
+    public float getY() {
+        return position.y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         position.y = y;
     }
 
-    public void setPosition(int x, int y) {
+    public void setPosition(float x, float y) {
         position.set(x, y);
     }
